true negative 100%..need to fix

+++ preprocessing/create_labels.py
@@ -1,23 +1,27 @@
 import numpy as np
 import pandas as pd
 import glob
-import sys
 import os
+from create_data_groups import fetch_files_with_numcalls
 
 
 def find_label(file, path):
     """
+    ***OLD:
     Output: Audit label file relating to input spectro file
     There are a few instances where a file has multiple audit label files:
     cc352a021_6h
     cc352b003_4h
     cc354a049_8h
-    in these cases the first label will be used.
+    in these cases the first label will be used.]
+
+    ***NEW:
+    Retrieving the label text file from the data set against a given file id
+    example: Return cc16_366a_344322s_labels.txt for audio file cc16_366a_344322s_audio.wav
     """
-    label_path = os.path.dirname(os.path.dirname(path)) + '/audit/'
-    label_search = ''.join(file.split('.')[0].split('_')[0]) + '_' + file.split('.')[0].split('_')[1]
-    labels_list = [name for name in glob.glob(label_path + label_search + '*')]
-    return labels_list[0]
+    label_search = file.rsplit("_",maxsplit=1)[0]
+    label = fetch_files_with_numcalls(path,1).loc[label_search]['labels']
+    return label
 
 
 def create_label_dataframe(label, begin_time, end_time, window_size, timesteps_per_second):
@@ -27,17 +31,21 @@ def create_label_dataframe(label, begin_time, end_time, window_size, timesteps_p
     """
     labels_df = pd.read_csv(label,
                             sep='\t',
-                            index_col='Selection')
+                            names=['StartTime','Duration','Label'])
+                            # index_col='Selection')
     if 'Label' in labels_df.columns:
         # filter for any labels that do not start with definitive call type label
-        call_labels = ['GIG', 'SQL', 'GRL', 'GRN', 'SQT', 'MOO', 'RUM', 'WHP']
+        # call_labels = ['GIG', 'SQL', 'GRL', 'GRN', 'SQT', 'MOO', 'RUM', 'WHP']
         # change labels to first 3 characters
         labels_df.Label = labels_df.Label.str[0:3]
-        labels_df = labels_df[labels_df['Label'].isin(call_labels)]
-        labels_df['Begin Time(t)'] = ((labels_df['Begin Time (s)'] - begin_time) * timesteps_per_second).apply(np.floor)
-        labels_df['End Time(t)'] = ((labels_df['End Time (s)'] - begin_time) * timesteps_per_second).apply(np.ceil)
-        labels_df = labels_df[labels_df['Begin Time (s)'] >= begin_time]
-        labels_df = labels_df[labels_df['End Time (s)'] <= end_time] 
+        # labels_df = labels_df[labels_df['Label'].isin(call_labels)]
+        labels_df['Begin Time(t)'] = ((labels_df['StartTime'] - begin_time) * timesteps_per_second).apply(np.floor)
+
+        labels_df['End Time(t)'] = ((labels_df['StartTime']+labels_df['Duration'] - begin_time) * timesteps_per_second).apply(np.floor)
+        labels_df = labels_df[labels_df['StartTime'] >= begin_time]
+        labels_df = labels_df[(labels_df['StartTime']+labels_df['Duration']) <= end_time]
+        if len(labels_df[labels_df['Label'].str.contains('\?')]) > 0:
+            labels_df.drop(labels_df.index, inplace=True)
     return labels_df
 
 
@@ -63,6 +71,8 @@ def create_label_matrix(dataframe, timesteps):
     [0, 0, 0, 0, 0, 0 ....],
     [1, 1, 1, 1, 0, 0 ....],] This represents a Whoop in 0-3 timesteps
     """
+    call_labels = ['GIG', 'SQL', 'GRL', 'GRN', 'SQT', 'MOO', 'RUM', 'WHP']
+    dataframe = dataframe[dataframe['Label'].isin(call_labels)]
     label = np.zeros((8, timesteps))
     if 'Label' in list(dataframe):
         # create update list
@@ -74,7 +84,7 @@ def create_label_matrix(dataframe, timesteps):
         # label correct row based on label
         for l in update_list:
             begin_t = int(l[0])
-            end_t = int(l[1])+1
+            end_t = int(l[1]) + 1
             if l[2] == 'GIG':
                 label[0][begin_t:end_t] = 1
             elif l[2] == 'SQL':
@@ -94,25 +104,27 @@ def create_label_matrix(dataframe, timesteps):
     return label
 
 
-paths = ['cc16_352a_converted/spectro/',
-	'cc16_352b_converted/spectro/',
-	'cc16_354a_converted/spectro/',
-	'cc16_360a_converted/spectro/',
-	'cc16_366a_converted/spectro/']
+paths = ['/cache/rmishra/cc16_366a_converted/spectro/']
+base_dir = "/cache/rmishra/cc16_ML/cc16_366a"
+
+
 
 for path in paths:
-	for f in os.listdir(path):
-	    if 'LABEL' not in f:
-	        label = find_label(f, path)
-	        begin_time = int(f.split('_')[2].split('sto')[0])
-	        end_time = int(f.split('_')[2].split('sto')[1].split('s')[0])
-	        window_size = end_time - begin_time
-	        timesteps = 259 # need to set timesteps
-	        timesteps_per_second = timesteps / window_size
-	        df = create_label_dataframe(label,
-	                                    begin_time,
-	                                    end_time,
-	                                    window_size,
-	                                    timesteps_per_second)
-	        label_matrix = create_label_matrix(df, timesteps)
-	        np.save(path+f[:-4]+'LABEL', label_matrix)
+    if(os.path.exists(path)):
+        for f in os.listdir(path):
+            if 'LABEL' not in f:
+                label = find_label(f, base_dir)           #We dont have different duration files
+                begin_time = int(f.split('_')[-1].split('sto')[0])
+                end_time = int(f.split('_')[-1].split('sto')[1].split('s')[0])
+                window_size = end_time - begin_time
+                timesteps = 259  # need to set timesteps
+                timesteps_per_second = timesteps / window_size
+                df = create_label_dataframe(os.path.join(base_dir,label),          #changing first parameter from label to f
+                                            begin_time,
+                                            end_time,
+                                            window_size,
+                                            timesteps_per_second)
+                label_matrix = create_label_matrix(df, timesteps)
+                print("Saving file: {}".format(path + f[:-4] + 'LABEL'))
+                np.save(path + f[:-4] + 'LABEL', label_matrix)
+                np.savetxt(path + f[:-4] + 'LABEL', label_matrix, delimiter=",")
